\name{moyenne()}
\alias{moyenne()}
\title{
  moyenne()
}
\description{
  Fonction Rcpp pour le F1 Project
}
\usage{
  moyenne(tab, taille_tab)
}
\arguments{
  \item{tab, taille_tab}{
      un vecteur d'entier et la taille du vecteur
}
}
\value{
  Retourne la moyenne du vecteur entré en paramètre
}
\author{
  Grancher Antoine, Champ Charline, Amistadi Batiste
}
